调用接口变为统一方法

+++ coc-api-master/coc-api-backend/src/main/java/com/lin/cocapibackend/controller/AnalysisController.java
@@ -51,7 +51,7 @@ public class AnalysisController {
         List<InterfaceInfoVO> interfaceInfoVOList = list.stream().map(interfaceInfo -> {
             InterfaceInfoVO interfaceInfoVO = new InterfaceInfoVO();
             BeanUtils.copyProperties(interfaceInfo, interfaceInfoVO);
-            int totalNum = interfaceInfoIdObjMap.get(interfaceInfo.getId()).get(0).getTotalNum();
+            Long totalNum = Long.valueOf(interfaceInfoIdObjMap.get(interfaceInfo.getId()).get(0).getTotalNum());
             interfaceInfoVO.setTotalNum(totalNum);
             return interfaceInfoVO;
         }).collect(Collectors.toList());
